e_learning_platform/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.4
djangorestframework==3.15.2
pillow==11.0.0
reportlab==4.2.5
sqlparse==0.5.3
